Adjustments
- Disable Admin > Connections page which shows S3 access keys in plain
  text
- Add run_ds, run_ts, run_date to task context
- Switch DAG processors to log to the scheduler stdout/stderr rather than
  separate files
- Reraise exceptions on object edit failure
- Override OAuth for Github with the right path, match only by email,
and add support for limiting to Github groups
- Run CompileAssets on build in order to populate the static directory,
also make build fail if the compile assets script failes (check_call)
- Fix dual Alembic migration heads
- Copy over all static assets in order to update the glyphicons font
which is 6 years old and missing the duplicate icon. For that to work we
also overwrite the static_folder of appbuilder
- Fix duration in tooltip, backporting https://github.com/apache/airflow/pull/5071
- Backport commit cb4e39a1ab903237cca6f9f965e5dd256a268491 from master,
bumping up Celery to 4.3.0. PR https://github.com/apache/airflow/pull/5116